Fixing deadlock when command prints a lot to stderr
Currently, only stdout is read on the fly. If the stderr pipe fills up,
then, the subprocess will deadlock when trying to write to stderr.
As the parent process, only reads to stdout, and waits for the process
to finish, this will never resolve. This uses instead the select
function to find which file descriptor can be read from.