Fixing deadlock when command prints a lot to stderr (#1609)
* Fixing deadlock when command prints a lot to stderr

Currently, only stdout is read on the fly. If the stderr pipe fills up,
then, the subprocess will deadlock when trying to write to stderr.
As the parent process, only reads to stdout, and waits for the process
to finish, this will never resolve. This uses instead the select
function to find which file descriptor can be read from.

* formatting

* fixing test compatibility python3.6

* refactoring to make linter happy

* making type system happy

* finally linter is happy

* extra comments

* use raw buffers

* better implementation

* make linter happy

* use executor for test

* specialize copy_streams code

* work on unbuffered io

Co-authored-by: Guillaume Wenzek <gwenzek@users.noreply.github.com>