Add helper class to facilitate PyTorch distributed initialization (#1703)
It is not as straightforward as it could be to initialize PyTorch distributed in a job scheduled with submitit. These changes introduce a helper class that maps the relevant job environment properties and export them as environment variables to easily setup PyTorch distributed with the (default) env:// initialization method.

Tested on a toy job (included as a complete example).